feat(panel): configurable output range on the lever
A lever could only ever emit a 0..15 redstone level, which makes it
useless for driving a value with its own scale — a camera FOV lives in
30..110, a motor setpoint somewhere else entirely. It now carries Min
and Max fields like the knob already did, edited with the Panel Key, and
its 16 detents map onto that range. The output pin becomes FLOAT so
fractional ranges work; with the default 0..15 it still reads exactly as
a redstone level, and the `set` input maps back the same way so remote
drive stays symmetric.

+++ src/main/kotlin/dev/nitka/nodewire/block/panel/PanelElementType.kt
@@ -49,7 +49,10 @@ object PanelElements {
         PanelElementType("key_switch", 2, 2, listOf(out("", PinType.BOOL))),
         // Cover is pin-driven (open), the big button is the output.
         PanelElementType("emergency", 4, 4, listOf(out("", PinType.BOOL), inp("open", PinType.BOOL))),
-        PanelElementType("lever", 3, 5, listOf(out("", PinType.INT), inp("set", PinType.INT))),
+        // Output is FLOAT because the 16 detents are mapped onto a configured
+        // min..max range (see PanelElementConfig); with the default 0..15 it
+        // still reads as a plain redstone level.
+        PanelElementType("lever", 3, 5, listOf(out("", PinType.FLOAT), inp("set", PinType.FLOAT))),
         PanelElementType("knob", 2, 2, listOf(out("", PinType.FLOAT), inp("set", PinType.FLOAT))),
         PanelElementType(
             "joystick", 4, 4,
